requirements.txt
# requirements.txt

Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
firebase_admin==7.1.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.5.0
mysqlclient==2.2.7
PyJWT==2.10.1
requests==2.32.5
gunicorn
whitenoise